refactor(slides): separate runtime and calibration ownership
